% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_funs.R
\name{calc_percent_repondents_by_var}
\alias{calc_percent_repondents_by_var}
\title{Calculate percent of respondents with contx_type X taxa_type by stratification var}
\usage{
calc_percent_repondents_by_var(contaxa_dat, demo_dat, var_name)
}
\arguments{
\item{contaxa_dat}{joined dataframe from join_contx_and_demo}

\item{demo_dat}{dataframe from select_demo_dat that was used in join_contx_and_demo function}

\item{var_name}{stratification variable name (character type)}
}
\value{
dataframe with percent of respondents normalized by stratification variable count
}
\description{
Calculate percent of respondents with contx_type X taxa_type by stratification var
}
