% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_funs.R
\name{calc_total_respondents}
\alias{calc_total_respondents}
\title{Return total number of respondents where var_name == var_val}
\usage{
calc_total_respondents(demo_dat, var_name, var_val)
}
\arguments{
\item{demo_dat}{from select_demo_dat function}

\item{var_name}{stratification variable name}

\item{var_val}{stratification variable value}
}
\description{
Return total number of respondents where var_name == var_val
}
\author{
matteo-V
}
