% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_funs.R
\name{expand_contaxa_by_var}
\alias{expand_contaxa_by_var}
\title{General expand contaxa by variable function}
\usage{
expand_contaxa_by_var(dat, var_name)
}
\arguments{
\item{dat}{dataframe returned from calc_percent_respondents_by_var function}

\item{var_name}{variable name used for stratification  as character}
}
\value{
dataframe with full crossproduct of (contx_type X taxa_type) filled in for viz functions
}
\description{
Fill in dataframe with all crossprods of (contx_type X taxa_type)
}
\author{
matteo-V
}
