% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz_funs.R
\name{get_age_ntile_factor_order}
\alias{get_age_ntile_factor_order}
\title{Get age quantile factor order}
\usage{
get_age_ntile_factor_order(dat, human_dat)
}
\arguments{
\item{human_dat}{EIDITH human data table, filtered for country of interest for plotting}
}
\value{
ordered character vector of age_quint_range levels
}
\description{
Get age quantile factor order
}
\author{
matteo-V
}
