% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_funs.R
\name{get_age_ntiles}
\alias{get_age_ntiles}
\title{Annotate dataframe with age quintiles}
\usage{
get_age_ntiles(dat, ntiles = 5)
}
\arguments{
\item{dat}{EIDITH demographic dataframe which to annotate (must contain age column)}

\item{ntiles}{passed to quntiles function}
}
\description{
Annotate dataframe with age quintiles
}
\author{
matteo-V
}
