% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_funs.R
\name{get_clean_illness_covariates}
\alias{get_clean_illness_covariates}
\title{Create and widen covariate data for self reported illness modeling}
\usage{
get_clean_illness_covariates(dat)
}
\arguments{
\item{dat}{a raw human table from EIDITH which to clean}
}
\value{
wide data set with data types cleaned for regression analysis
}
\description{
Create and widen covariate data for self reported illness modeling
}
\author{
matteo-V
}
