% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_funs.R
\name{get_clean_occupations}
\alias{get_clean_occupations}
\title{Create new variables occupation using all pertinent data from survey, can also aggregate rare}
\usage{
get_clean_occupations(human_dat, aggregate_rare = T, threshold = 10,
  aggregate_value = "rare")
}
\arguments{
\item{human_dat}{EIDITH human data table}

\item{threshold}{number below which a occupation is considered 'rare'}

\item{aggregateRate}{logical indicating whether uncommon occupations should be aggregated}
}
\value{
human data table with new occupation column
}
\description{
Create new variables occupation using all pertinent data from survey, can also aggregate rare
}
\author{
matteo-V
}
