% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_funs.R
\name{get_concurrent_sites}
\alias{get_concurrent_sites}
\title{Get concurrent site names for within country visualizations}
\usage{
get_concurrent_sites(human_dat)
}
\arguments{
\item{human_dat}{from ed2_human}

\item{event_dat}{from ed2_event}
}
\value{
human_dat frame with concurrent sites added
}
\description{
Get concurrent site names for within country visualizations
}
\author{
matteo-V
}
