% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_funs.R
\name{get_country_codes}
\alias{get_country_codes}
\title{Create varaible with two letter country code for cross country viz}
\usage{
get_country_codes(human_dat)
}
\arguments{
\item{human_dat}{EIDITH human table with event_name}
}
\value{
human_dat frame with country code added
}
\description{
Create varaible with two letter country code for cross country viz
}
\author{
matteo-V
}
