% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz_funs.R
\name{get_education_factor_order}
\alias{get_education_factor_order}
\title{Get education factor order}
\usage{
get_education_factor_order(dat, strat_var)
}
\arguments{
\item{dat}{dataframe with education factor levels (with annotation) to reorder}

\item{strat_var}{stratification variable}
}
\value{
vector of factor levels ordered by decreasing education
}
\description{
Get education factor order
}
\author{
matteo-V
}
