% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_funs.R
\name{get_illness_analysis_dat}
\alias{get_illness_analysis_dat}
\title{Create illness analysis frame}
\usage{
get_illness_analysis_dat(dat, outcome_var, taxa_names = c("rodents", "nhp",
  "bats", "swine", "poultry"))
}
\arguments{
\item{dat}{data frame from which to create analysis data set created from get_self_reports or get_contaxa}

\item{outcome_var}{outcome variable of interest keyword (sari, ili, encehpalitis or hemorrhagic)}

\item{taxa_names}{vector of taxa types for which exposures willbe included}
}
\description{
Creates data frame ready for random forest
}
\author{
matteo-V
}
