% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_funs.R
\name{get_self_reports}
\alias{get_self_reports}
\title{Obtain and flatten self reported symptom outcome data}
\usage{
get_self_reports(dat, of_interest = T)
}
\arguments{
\item{dat}{data frame for which to obtain self-reported outcomes}

\item{of_interest}{logical indicating to only select outcomes of interest as per analysis plan}
}
\value{
data frame (tibble) with self reported illness widened to dichotomous variables
}
\description{
Obtain and flatten self reported symptom outcome data
}
\author{
matteo-V
}
