% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_funs.R
\name{join_contx_and_demo}
\alias{join_contx_and_demo}
\title{Join contx data and demo data}
\usage{
join_contx_and_demo(contx_dat, demo_dat)
}
\arguments{
\item{contx_dat}{from reshape_contaxa_data function}

\item{demo_dat}{from select_demo_dat function}
}
\value{
joined contaxa  and demographic variables frames
}
\description{
Join contx data and demo data
}
