% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz_funs.R
\name{plot_contact_wrap_heatmap}
\alias{plot_contact_wrap_heatmap}
\title{Plot heatmap wrapped by contact type}
\usage{
plot_contact_wrap_heatmap(contaxa_dat, strat_var, factor_order = NULL)
}
\arguments{
\item{strat_var}{varaible to use on y axis}

\item{factor_order}{used by plot_contaxa_heatmap to order plotting factors}
}
\value{
unstyled heatmap to be faceted by plot_contaxa_heatmap
}
\description{
Plot heatmap wrapped by contact type
}
\author{
matteo-V
}
