% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz_funs.R
\name{plot_contaxa_heatmap}
\alias{plot_contaxa_heatmap}
\title{Main heatmap function}
\usage{
plot_contaxa_heatmap(dat, strat, human_dat = NULL, wrap_contact = T)
}
\arguments{
\item{dat}{contaxa dataframe of frequencies by stratification variable}

\item{strat}{string of variable name for stratification}

\item{human_dat}{(optional) used for ordering plots by age_quinte_range}
}
\value{
ggplot2 heatmap object
}
\description{
Main heatmap function
}
\author{
matteo-V
}
