% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz_funs.R
\name{plot_strat_wrap_heatmap}
\alias{plot_strat_wrap_heatmap}
\title{Create heat map of contx_type X taxa_type}
\usage{
plot_strat_wrap_heatmap(contaxa_dat)
}
\arguments{
\item{contaxa_data}{data passed to}
}
\value{
heatmap to be faceted by plot_contaxa_heatmap
}
\description{
Create heat map of contx_type X taxa_type
}
\author{
matteo-V
}
