% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_funs.R
\name{reshape_contaxa_data}
\alias{reshape_contaxa_data}
\title{Unnest and restructure taxa contact cols into}
\usage{
reshape_contaxa_data(dat, taxa)
}
\arguments{
\item{dat}{returned from select_taxa_dat function}

\item{taxa}{vector of taxa types of interest}
}
\value{
dataframe with participant_id (key), taxa_type, contx_type, and aggregate sum
}
\description{
Unnest and restructure taxa contact cols into
}
\author{
: matteo-V
}
