% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_funs.R
\name{run_contaxa_crosstab}
\alias{run_contaxa_crosstab}
\title{Run crosstab for exposures (taxa contacts)}
\usage{
run_contaxa_crosstab(analysis_dat, condition_var, alpha = 0.05)
}
\arguments{
\item{analysis_dat}{cleaned covariate data}

\item{condition_var}{variable on which to condition (test) for the exposures}

\item{alpha}{the alpha value for which to assess significance}
}
\value{
data frame of results for condition, taxa, contact type, p-value and significance (bonferroni corrected)
}
\description{
Run crosstab for exposures (taxa contacts)
}
\author{
matteo-V
}
