% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_funs.R
\name{run_crosstab}
\alias{run_crosstab}
\title{Crosstab, run fisher test (robust to small counts) and output p-value}
\usage{
run_crosstab(dat, condition, outcome)
}
\arguments{
\item{condition}{for which to test independence}

\item{outcome}{the outcome variable of interest}
}
\value{
p value of fisher test
}
\description{
Crosstab, run fisher test (robust to small counts) and output p-value
}
\author{
matteo-V
}
