% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_funs.R
\name{select_country_dat}
\alias{select_country_dat}
\title{Filter country data by country codes}
\usage{
select_country_dat(dat, country_codes)
}
\arguments{
\item{dat}{human data table from EIDITH}

\item{country_codes}{a vector of 2 digit country codes of interest}
}
\value{
full data set, filtered by countries of interest
}
\description{
Filter country data by country codes
}
\author{
matteo-V
}
