% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_funs.R
\name{select_demo_dat}
\alias{select_demo_dat}
\title{Select demographic vars from data}
\usage{
select_demo_dat(dat)
}
\arguments{
\item{dat}{EIDITH human dataframe from which to select variables}
}
\value{
dataframe of demographic data and participantID (key)
}
\description{
Select demographic vars from data
}
\author{
matteo-V
}
