% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip_funs.R
\name{select_taxa_dat}
\alias{select_taxa_dat}
\title{Select taxa contact cols by taxa names}
\usage{
select_taxa_dat(human_dat, taxa_types)
}
\arguments{
\item{taxa_types}{a vector of taxa names of interest}

\item{dat}{is the data set from which to select taxa columns}
}
\value{
participantID (key) and taxa columns of interest
}
\description{
Select taxa contact cols by taxa names
}
\author{
matteo-V
}
