% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_funs.R
\name{train_test_split}
\alias{train_test_split}
\title{Train test val split function}
\usage{
train_test_split(dat)
}
\arguments{
\item{dat}{dataframe to split for training}
}
\value{
named list of data frames
}
\description{
Train test val split function
}
\author{
matteo-V
}
